+++ src/main/java/io/cavia/trader/module/member/mypage/service/MyPageServiceImpl.java
@@ -57,7 +57,7 @@ public class MyPageServiceImpl implements MyPageService {
     }
 
     @Override
-    public int changePassword(int id, String password, LocalDateTime passwordUpdatedAt) {
+    public int changePassword(Long id, String password, LocalDateTime passwordUpdatedAt) {
         return mybatisMemberRepository.updatePassword(id,password,passwordUpdatedAt);
     }
