+++ src/main/java/io/cavia/trader/module/member/repository/MybatisMemberRepository.java
@@ -24,13 +24,19 @@ public class MybatisMemberRepository implements MemberRepository {
     }
 
     @Override
-    public int updateNickname(int id, String nickname, LocalDateTime nicknameUpdatedAt) {return memberMapper.updateNickname(id,nickname,nicknameUpdatedAt);}
+    public int updateNickname(int id, String nickname, LocalDateTime nicknameUpdatedAt) {
+        return memberMapper.updateNickname(id, nickname, nicknameUpdatedAt);
+    }
 
     @Override
-    public int updatePassword(int id, String password, LocalDateTime passwordUpdatedAt) {return memberMapper.updatePassword(id,password,passwordUpdatedAt);}
+    public int updatePassword(Long id, String password, LocalDateTime passwordUpdatedAt) {
+        return memberMapper.updatePassword(id, password, passwordUpdatedAt);
+    }
 
     @Override
-    public int updateCash(int id, Long cash) {return memberMapper.updateCash(id,cash);}
+    public int updateCash(int id, Long cash) {
+        return memberMapper.updateCash(id, cash);
+    }
 
     @Override
     public boolean existsByEmail(String email) {
@@ -43,7 +49,9 @@ public class MybatisMemberRepository implements MemberRepository {
     }
 
     @Override
-    public boolean existsByIdAndPassword(int id,String password) {return memberMapper.existsByIdAndPassword(id,password);}
+    public boolean existsByIdAndPassword(int id, String password) {
+        return memberMapper.existsByIdAndPassword(id, password);
+    }
 
     @Override
     public Optional<Member> findById(Long id) {
